Update task 5 #8

+++ 0x0C-more_malloc_free/101-mul.c
@@ -118,7 +118,7 @@ int main(int argc, char *argv[])
 	}
 	else
 	{
-		printf("%d\n", (atoi(argv[1]) * atoi(argv[2])));
+		printf("%lu\n", (unsigned long)((atoi(argv[1]) * atoi(argv[2]))));
 	}
 	return (0);
 }
